+++ ibmsecurity/isam/base/ssl_certificates/certificate_databases.py
@@ -1,8 +1,11 @@
 import logging
 import zipfile
 import shutil
+import json
+# from codecs import ignore_errors
 
 from ibmsecurity.utilities.tools import get_random_temp_dir
+from ibmsecurity.utilities.tools import json_equals
 
 logger = logging.getLogger(__name__)
 requires_model = "Appliance"
@@ -49,13 +52,19 @@ def get(isamAppliance, cert_dbase_id, check_mode=False, force=False):
     """
     Retrieving the SSL certificate database details
     """
-    return isamAppliance.invoke_get("Retrieving the SSL certificate database details",
+
+    retObj = isamAppliance.invoke_get("Retrieving the SSL certificate database details",
                                     f"/isam/ssl_certificates/{cert_dbase_id}/details",
-                                    requires_model=requires_model)
+                                    requires_model=requires_model,
+                                    ignore_error=True)
+    if retObj.get('rc', 0) == 404:
+        return isamAppliance.create_return_object(rc=404, warnings=[f"{cert_dbase_id} does not exist"])
+    else:
+        return retObj
 
 
 def create(isamAppliance, kdb_name, type='kdb',
-           check_mode=False, force=False
+           check_mode=False, force=False,
            **kwargs,
           ):
     """
@@ -74,10 +83,9 @@ def create(isamAppliance, kdb_name, type='kdb',
     """
     warnings = []
     if force or not _check(isamAppliance, kdb_name):
-        if check_modee:
+        if check_mode:
             return isamAppliance.create_return_object(changed=True)
         else:
-
             json_data = {
                     "kdb_name": kdb_name,
                     "type": type,
@@ -87,19 +95,19 @@ def create(isamAppliance, kdb_name, type='kdb',
                     json_data[k] = v
                # new in 10.0.5
                if k in certificate_database_new_10_05:
-                   if ibmsecurity.utilities.tools.version_compare(isamAppliance.facts['version'], "10.0.8.0") < 0:
-                       warnings.append(                         f"Appliance at version: {isamAppliance.facts['version']}, load_certificate: {load_certificate} is not supported. Needs 10.0.0.0 or higher. Ignoring load_certificate for this call.")
+                   if ibmsecurity.utilities.tools.version_compare(isamAppliance.facts['version'], "10.0.5.0") < 0:
+                       warnings.append(f"Appliance at version: {isamAppliance.facts['version']}, {k}: {v} is not supported. Needs 10.0.5.0 or higher. Ignoring {k} for this call.")
                    else:
                        json_data[k] = v
                # new in 10.0.8
                if k in certificate_database_new_10_08:
                    if ibmsecurity.utilities.tools.version_compare(isamAppliance.facts['version'], "10.0.8.0") < 0:
-                       warnings.append(                         f"Appliance at version: {isamAppliance.facts['version']}, load_certificate: {load_certificate} is not supported. Needs 10.0.0.0 or higher. Ignoring load_certificate for this call.")
+                       warnings.append(f"Appliance at version: {isamAppliance.facts['version']}, {k}: {v} is not supported. Needs 10.0.8.0 or higher. Ignoring {k} for this call.")
                    else:
                        json_data[k] = v
                # Logic
                if type != 'p11' or json_data.get("hsm_type", None) not in ("safenet","safenet-ha"):
-                   warnings.append("Serial number, safenet_* are only valid for safenet hsm, removed from input")
+                   # warnings.append("Serial number, safenet_* are only valid for safenet hsm, removed from input")
                    json_data.pop("serial_number", None)
                    json_data.pop("safenet_user", None)
                    json_data.pop("safenet_pw", None)
@@ -118,12 +126,17 @@ def create(isamAppliance, kdb_name, type='kdb',
                    json_data.pop("rfs_port", None)
                    json_data.pop("rfs_auth", None)
 
-            return isamAppliance.invoke_post(
-                f"Creating certificate database '{kdb_name}'",
-                "/isam/ssl_certificates",
-                json_data,
-                warnings=warnings
-            )
+               retObj = isamAppliance.invoke_post( f"Creating certificate database {kdb_name}",
+                  "/isam/ssl_certificates",
+                  json_data,
+                  warnings=warnings,
+                  ignore_error=True
+               )
+               if retObj.get("rc", 0) == 400:
+                   warnings.append(f"Invalid type (you need to install an extension to support network hsm {type})")
+                   return isamAppliance.create_return_object(warnings=warnings)
+               else:
+                   return retObj
 
     return isamAppliance.create_return_object()
 
@@ -222,31 +235,66 @@ def rename(isamAppliance, cert_id, new_name, check_mode=False, force=False):
     return isamAppliance.create_return_object()
 
 
-def set(isamAppliance, cert_id, description, check_mode=False, force=False):
+def set(isamAppliance, cert_id, description=None, type="kdb", check_mode=False, force=False, **kwargs):
     """
-    Set description for a certificate database
+    Set description for a certificate database (type="kdb)"
+    Update network certificate database (type="p11")
     """
+    warnings = []
     desc_match = True  # This will remain True even when cert db is not found!
-    if force is False:
-        ret_obj = get_all(isamAppliance)
-        for certdb in ret_obj['data']:
-            if certdb['id'] == cert_id:
-                if certdb['description'] != description:
-                    desc_match = False
-                break
 
-    if force is True or desc_match is False:
-        if check_mode is True:
-            return isamAppliance.create_return_object(changed=True)
-        else:
-            return isamAppliance.invoke_put(
-                "Set description for a certificate database",
-                f"/isam/ssl_certificates/{cert_id}",
-                {
-                    "description": description
-                })
+    if type == "kdb":
+        if not force:
+            if description is None:
+                desc_match = True
+            else:
+                ret_obj = get_all(isamAppliance)
+                for certdb in ret_obj['data']:
+                    if certdb['id'] == cert_id:
+                        if certdb['description'] != description:
+                            desc_match = False
+                        break
 
-    return isamAppliance.create_return_object()
+        if force or not desc_match:
+            if check_mode is True:
+                return isamAppliance.create_return_object(changed=True)
+            else:
+                return isamAppliance.invoke_put(
+                    "Set description for a certificate database",
+                    f"/isam/ssl_certificates/{cert_id}",
+                    {
+                        "description": description
+                    })
+    else:
+        # type = p11
+        json_data = {}
+        for k, v in kwargs.items():
+            if k in certificate_database_optional_args:
+                json_data[k] = v
+            # new in 10.0.5
+            if k in certificate_database_new_10_05:
+                if ibmsecurity.utilities.tools.version_compare(isamAppliance.facts['version'], "10.0.5.0") < 0:
+                    warnings.append(f"Appliance at version: {isamAppliance.facts['version']}, {k}: {v} is not supported. Needs 10.0.5.0 or higher. Ignoring {k} for this call.")
+                else:
+                    json_data[k] = v
+            # new in 10.0.8
+            if k in certificate_database_new_10_08:
+                if ibmsecurity.utilities.tools.version_compare(isamAppliance.facts['version'], "10.0.8.0") < 0:
+                    warnings.append(f"Appliance at version: {isamAppliance.facts['version']}, {k}: {v} is not supported. Needs 10.0.8.0 or higher. Ignoring {k} for this call.")
+                else:
+                    json_data[k] = v
+        ret_obj = get(isamAppliance, cert_id)
+        # Idempotency
+        if ret_obj.get("rc", 0) == 404:
+            return isamAppliance.create_return_object(warnings=warnings)
+        if not json_equals(ret_obj, json_data):
+                return isamAppliance.invoke_put(
+                    "Updating network certificate database",
+                    f"/isam/ssl_certificates/{cert_id}",
+                    json_data,
+                    warnings=warnings
+                )
+    return isamAppliance.create_return_object(warnings=warnings)
 
 
 def _check(isamAppliance, id):