+++ ibmsecurity/isam/statistics.py
@@ -9,9 +9,7 @@ def get_network(isamAppliance, application_interface, statistics_duration, check
     Retrieving the Application Interface Statistics
     """
     return isamAppliance.invoke_get("Retrieving the Application Interface Statistics",
-                                    "/analysis/interface_statistics.json{0}".format(
-                                        tools.create_query_string(prefix=application_interface,
-                                                                  timespan=statistics_duration)),requires_model=requires_model)
+                                    f"/analysis/interface_statistics.json{tools.create_query_string(prefix=application_interface, timespan=statistics_duration)}",requires_model=requires_model)
 
 
 def get_rp_junction(isamAppliance, instance, date, duration, check_mode=False, force=False):
@@ -19,10 +17,7 @@ def get_rp_junction(isamAppliance, instance, date, duration, check_mode=False, f
     Retrieving junction average response times for a Reverse Proxy instance
     """
     return isamAppliance.invoke_get("Retrieving junction average response times for a Reverse Proxy instance",
-                                    "/analysis/reverse_proxy_traffic/reqtime{0}".format(
-                                        tools.create_query_string(date=date,
-                                                                  duration=duration,
-                                                                  instance=instance)),requires_model=requires_model)
+                                    f"/analysis/reverse_proxy_traffic/reqtime{tools.create_query_string(date=date, duration=duration, instance=instance)}",requires_model=requires_model)
 
 
 def get_rp_health_summary(isamAppliance, check_mode=False, force=False):
@@ -39,11 +34,7 @@ def get_rp_throughput_summary(isamAppliance, date, duration, aspect, summary=Non
     """
     headers = {'Accept': 'application/json', 'range': 'items=0-24'}
     return isamAppliance.invoke_get_with_headers("Retrieving a summary of throughput for all Reverse Proxy instances",
-                                                 "/analysis/reverse_proxy_traffic/throughput/{0}".format(
-                                                     tools.create_query_string(summary=summary,
-                                                                               date=date,
-                                                                               duration=duration,
-                                                                               aspect=aspect)),
+                                                 f"/analysis/reverse_proxy_traffic/throughput/{tools.create_query_string(summary=summary, date=date, duration=duration, aspect=aspect)}",
                                                                                requires_model=requires_model,
                                                                                headers=headers)
 
@@ -53,10 +44,7 @@ def get_rp_throughput(isamAppliance, instance, date, duration, check_mode=False,
     Retrieving throughput records for a specific Reverse Proxy instance
     """
     return isamAppliance.invoke_get("Retrieving throughput records for a specific Reverse Proxy instance",
-                                    "/analysis/reverse_proxy_traffic/throughput/{0}{1}".format(instance,
-                                                                                               tools.create_query_string(
-                                                                                                   date=date,
-                                                                                                   duration=duration)),requires_model=requires_model)
+                                    f"/analysis/reverse_proxy_traffic/throughput/{instance}{tools.create_query_string(date=date, duration=duration)}",requires_model=requires_model)
 
 
 def get_rp_traffic_summary(isamAppliance, instance, date, duration, aspect, summary=True, check_mode=False,
@@ -66,11 +54,7 @@ def get_rp_traffic_summary(isamAppliance, instance, date, duration, aspect, summ
     """
     return isamAppliance.invoke_get(
         "Retrieving a summary of traffic by Junction or User-Agent on a Reverse Proxy instance",
-        "/analysis/reverse_proxy_traffic/traffic/instance/{0}/{1}".format(instance,
-                                                                          tools.create_query_string(summary=summary,
-                                                                                                    date=date,
-                                                                                                    duration=duration,
-                                                                                                    aspect=aspect)),requires_model=requires_model)
+        f"/analysis/reverse_proxy_traffic/traffic/instance/{instance}/{tools.create_query_string(summary=summary, date=date, duration=duration, aspect=aspect)}",requires_model=requires_model)
 
 
 def get_rp_traffic(isamAppliance, instance, date, duration, aspect, aspect_identifier, check_mode=False, force=False):
@@ -79,9 +63,7 @@ def get_rp_traffic(isamAppliance, instance, date, duration, aspect, aspect_ident
     """
     return isamAppliance.invoke_get(
         "Retrieving a summary of traffic records for a specific Junction or User-Agent on a Reverse Proxy instance",
-        "/analysis/reverse_proxy_traffic/traffic/instance/{0}/{1}/{2}{3}".format(instance, aspect, aspect_identifier,
-                                                                                 tools.create_query_string(date=date,
-                                                                                                           duration=duration)),requires_model=requires_model)
+        f"/analysis/reverse_proxy_traffic/traffic/instance/{instance}/{aspect}/{aspect_identifier}{tools.create_query_string(date=date, duration=duration)}",requires_model=requires_model)
 
 
 def get_rp_traffic_detail(isamAppliance, instance, date, duration, aspect, aspect_identifier, check_mode=False,
@@ -91,10 +73,7 @@ def get_rp_traffic_detail(isamAppliance, instance, date, duration, aspect, aspec
     """
     return isamAppliance.invoke_get(
         "Retrieving detailed traffic records for a specific Junction or User-Agent on a Reverse Proxy instance",
-        "/analysis/reverse_proxy_traffic/traffic/instance/{0}/{1}/{2}/{3}".format(instance, aspect, aspect_identifier,
-                                                                                  tools.create_query_string(date=date,
-                                                                                                            duration=duration,
-                                                                                                            aspect=aspect)),requires_model=requires_model)
+        f"/analysis/reverse_proxy_traffic/traffic/instance/{instance}/{aspect}/{aspect_identifier}/{tools.create_query_string(date=date, duration=duration, aspect=aspect)}",requires_model=requires_model)
 
 
 def get_rp_traffic_detail_aspect(isamAppliance, instance, date, duration, aspect, aspect_identifier, check_mode=False,
@@ -104,11 +83,7 @@ def get_rp_traffic_detail_aspect(isamAppliance, instance, date, duration, aspect
     """
     return isamAppliance.invoke_get(
         "Retrieving detailed traffic records for a specific User-Agent on a specific junction in a Reverse Proxy instance",
-        "/analysis/reverse_proxy_traffic/traffic/instance/{0}/{1}/{2}/{1}/{2}{3}".format(instance, aspect,
-                                                                                         aspect_identifier,
-                                                                                         tools.create_query_string(
-                                                                                             date=date,
-                                                                                             duration=duration)),requires_model=requires_model)
+        f"/analysis/reverse_proxy_traffic/traffic/instance/{instance}/{aspect}/{aspect_identifier}/{aspect}/{aspect_identifier}{tools.create_query_string(date=date, duration=duration)}",requires_model=requires_model)
 
 
 def get_rp_waf_events(isamAppliance, instance, date, duration, type, check_mode=False,
@@ -118,12 +93,7 @@ def get_rp_waf_events(isamAppliance, instance, date, duration, type, check_mode=
     """
     return isamAppliance.invoke_get(
         "Retrieving security action events for a Reverse Proxy instance",
-        "/analysis/reverse_proxy_traffic/pam_events{0}".format(
-            tools.create_query_string(
-                date=date,
-                duration=duration,
-                instance=instance,
-                type=type)),requires_model=requires_model)
+        f"/analysis/reverse_proxy_traffic/pam_events{tools.create_query_string(date=date, duration=duration, instance=instance, type=type)}",requires_model=requires_model)
 
 
 def get_cpu(isamAppliance, statistics_duration, check_mode=False, force=False):
@@ -132,9 +102,7 @@ def get_cpu(isamAppliance, statistics_duration, check_mode=False, force=False):
     """
     return isamAppliance.invoke_get(
         "Retrieving the CPU Usage Statistics",
-        "/statistics/systems/cpu.json{0}".format(
-            tools.create_query_string(
-                timespan=statistics_duration)),requires_model=requires_model)
+        f"/statistics/systems/cpu.json{tools.create_query_string(timespan=statistics_duration)}",requires_model=requires_model)
 
 
 def get_memory(isamAppliance, statistics_duration, check_mode=False, force=False):
@@ -143,9 +111,7 @@ def get_memory(isamAppliance, statistics_duration, check_mode=False, force=False
     """
     return isamAppliance.invoke_get(
         "Retrieving the Memory Usage Statistics",
-        "/statistics/systems/memory.json{0}".format(
-            tools.create_query_string(
-                timespan=statistics_duration)),requires_model=requires_model)
+        f"/statistics/systems/memory.json{tools.create_query_string(timespan=statistics_duration)}",requires_model=requires_model)
 
 
 def get_storage(isamAppliance, statistics_duration, check_mode=False, force=False):
@@ -154,6 +120,4 @@ def get_storage(isamAppliance, statistics_duration, check_mode=False, force=Fals
     """
     return isamAppliance.invoke_get(
         "Retrieving the Storage Usage Statistics",
-        "/statistics/systems/storage.json{0}".format(
-            tools.create_query_string(
-                timespan=statistics_duration)),requires_model=requires_model)
+        f"/statistics/systems/storage.json{tools.create_query_string(timespan=statistics_duration)}",requires_model=requires_model)