+++ ibmsecurity/isam/web/kerberos_configuration/keyfiles.py
@@ -26,11 +26,11 @@ def _check(isamAppliance, id):
     """
     ret_obj = get(isamAppliance)
 
-    logger.debug("Looking for {0} existing keytab files in: {1}".format(id, ret_obj['data']))
+    logger.debug(f"Looking for {id} existing keytab files in: {ret_obj['data']}")
     if ret_obj['data']:
         for keytab in ret_obj['data']:
             if keytab['id'] == id:
-                logger.debug("Found keytab: {0}".format(id))
+                logger.debug(f"Found keytab: {id}")
                 return True
 
     return False
@@ -40,12 +40,12 @@ def import_keytab(isamAppliance, id, file, check_mode=False, force=False):
     """
     Import a keytab file, id is the name of the keytab file (they should be the same)
     """
-    if force is True or _check(isamAppliance, id) is False:
-        if check_mode is True:
+    if force or not _check(isamAppliance, id):
+        if check_mode:
             return isamAppliance.create_return_object(changed=True)
         else:
             return isamAppliance.invoke_post_files(description="Import a keytab file",
-                                                   uri="{0}".format(uri),
+                                                   uri=uri,
                                                    fileinfo=[{
                                                        'file_formfield': 'keytab_file',
                                                        'filename': file,
@@ -58,15 +58,42 @@ def import_keytab(isamAppliance, id, file, check_mode=False, force=False):
     return isamAppliance.create_return_object()
 
 
+def export_keytab(isamAppliance, id, file, check_mode=False, force=False):
+    """
+    Export a keytab file, id is the name of the keytab file (they should be the same)
+    """
+    warnings=[]
+    if force or _check(isamAppliance, id):
+        if check_mode:
+            return isamAppliance.create_return_object(changed=True)
+        else:
+            retObj = isamAppliance.invoke_get_file(description="Export keytab file",
+                                                 uri=f"{uri}/{id}",
+                                                 filename=file,
+                                                 requires_modules=requires_modules,
+                                                 requires_version=requires_version,
+                                                 ignore_error=True
+                                                 )
+            if retObj.get("rc", 0) == 404:
+                logger.info(f"No keytab found matching your arguments {id}")
+                warnings.append(f"No keytab found matching your arguments {id}")
+                return isamAppliance.create_return_object(warnings=warnings)
+            else:
+                return retObj
+    warnings.append(f"No keytab found matching your arguments {id}")
+    return isamAppliance.create_return_object(warnings=warnings)
+
+
 def delete(isamAppliance, id, check_mode=False, force=False):
     """
     Delete a keytab file, id is the name of the keytab file
     """
-    if force is True or _check(isamAppliance, id) is True:
-        if check_mode is True:
+    if force or _check(isamAppliance, id):
+        if check_mode:
             return isamAppliance.create_return_object(changed=True)
         else:
-            return isamAppliance.invoke_delete(description="Delete a keytab file", uri="{0}/{1}".format(uri, id),
+            return isamAppliance.invoke_delete(description="Delete a keytab file",
+                                               uri=f"{uri}/{id}",
                                                requires_modules=requires_modules, requires_version=requires_version)
 
     return isamAppliance.create_return_object()
@@ -81,11 +108,11 @@ def combine(isamAppliance, newname, keytab_files, check_mode=False, force=False)
     """
     warnings = []
     for keytab in keytab_files:
-        if _check(isamAppliance, keytab) is False:
-            warnings.append("keytab file to be combined: {0}, not found".format(keytab))
+        if not _check(isamAppliance, keytab):
+            warnings.append(f"keytab file to be combined: {keytab}, not found")
 
-    if force is True or _check(isamAppliance, newname) is False:
-        if check_mode is True:
+    if force or not _check(isamAppliance, newname):
+        if check_mode:
             return isamAppliance.create_return_object(changed=True)
         else:
             return isamAppliance.invoke_put(description="Combine keytab files", uri=uri,